lab5

+++ src/Controller/CityController.php
@@ -16,6 +16,7 @@ class CityController extends AbstractController
     #[Route('/', name: 'app_city_index', methods: ['GET'])]
     public function index(CityRepository $cityRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_CITY_INDEX');
         return $this->render('city/index.html.twig', [
             'cities' => $cityRepository->findAll(),
         ]);
@@ -24,6 +25,7 @@ class CityController extends AbstractController
     #[Route('/new', name: 'app_city_new', methods: ['GET', 'POST'])]
     public function new(Request $request, CityRepository $cityRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_CITY_NEW');
         $city = new City();
         $form = $this->createForm(CityType::class, $city, [
 			'validation_groups' => ['new']]);
@@ -44,6 +46,7 @@ class CityController extends AbstractController
     #[Route('/{id}', name: 'app_city_show', methods: ['GET'])]
     public function show(City $city): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_CITY_SHOW');
         return $this->render('city/show.html.twig', [
             'city' => $city,
         ]);
@@ -52,6 +55,7 @@ class CityController extends AbstractController
     #[Route('/{id}/edit', name: 'app_city_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, City $city, CityRepository $cityRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_CITY_EDIT');
         $form = $this->createForm(CityType::class, $city, [
 			'validation_groups' => ['edit']]);
         $form->handleRequest($request);
@@ -71,6 +75,7 @@ class CityController extends AbstractController
     #[Route('/{id}', name: 'app_city_delete', methods: ['POST'])]
     public function delete(Request $request, City $city, CityRepository $cityRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_CITY_DELETE');
         if ($this->isCsrfTokenValid('delete'.$city->getId(), $request->request->get('_token'))) {
             $cityRepository->remove($city, true);
         }
