+++ src/Controller/WeatherCrudController.php
@@ -16,6 +16,7 @@ class WeatherCrudController extends AbstractController
     #[Route('/', name: 'app_weather_crud_index', methods: ['GET'])]
     public function index(WeatherRepository $weatherRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_WEATHER_CRUD_INDEX');
         return $this->render('weather_crud/index.html.twig', [
             'weather' => $weatherRepository->findAll(),
         ]);
@@ -24,6 +25,7 @@ class WeatherCrudController extends AbstractController
     #[Route('/new', name: 'app_weather_crud_new', methods: ['GET', 'POST'])]
     public function new(Request $request, WeatherRepository $weatherRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_WEATHER_CRUD_NEW');
         $weather = new Weather();
         $form = $this->createForm(WeatherType::class, $weather, [
 			'validation_groups' => ['new']]);
@@ -44,6 +46,7 @@ class WeatherCrudController extends AbstractController
     #[Route('/{id}', name: 'app_weather_crud_show', methods: ['GET'])]
     public function show(Weather $weather): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_WEATHER_CRUD_SHOW');
         return $this->render('weather_crud/show.html.twig', [
             'weather' => $weather,
         ]);
@@ -52,6 +55,7 @@ class WeatherCrudController extends AbstractController
     #[Route('/{id}/edit', name: 'app_weather_crud_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Weather $weather, WeatherRepository $weatherRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_WEATHER_CRUD_EDIT');
         $form = $this->createForm(WeatherType::class, $weather, [
 			'validation_groups' => ['edit']]);
         $form->handleRequest($request);
@@ -71,6 +75,7 @@ class WeatherCrudController extends AbstractController
     #[Route('/{id}', name: 'app_weather_crud_delete', methods: ['POST'])]
     public function delete(Request $request, Weather $weather, WeatherRepository $weatherRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_WEATHER_CRUD_DELETE');
         if ($this->isCsrfTokenValid('delete'.$weather->getId(), $request->request->get('_token'))) {
             $weatherRepository->remove($weather);
         }
